JPIE : Correction du module de mise à jour et de l'extract zeybux.

+++ branches/1.0/Maintenance/Maj/deploiementFile.php
@@ -5,11 +5,11 @@
 function supprimerDossier($pPath) {
 	$d = dir($pPath);
 	while (false !== ($entry = $d->read())) {	   
-	   if(	$entry != '.' && $entry != '..' && $entry != 'index.html' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'logs' && $entry != "DB.php" && $entry != "Mail.php"
+	   if(	$entry != '.' && $entry != '..' && $d->path.'/'.$entry != '../index.html' && $entry != 'Maintenance.php' && $entry != 'Maintenance' && $entry != 'logs' && $entry != "DB.php" && $entry != "Mail.php"
 	    && $entry != ".htaccess") {
    		if(is_dir($d->path.'/'.$entry)) {
    			supprimerDossier($d->path.'/'.$entry);
-			if($entry != 'configuration') {
+			if($entry != 'configuration' && $entry != 'classes' && $entry != 'html' && $entry != 'vues') {
 				rmdir($d->path.'/'.$entry);
 			}
    		} else {
@@ -26,7 +26,7 @@ supprimerDossier(DOSSIER_SITE);
 function parcourirDossier($pPathIn,$pPathOut) {
 	$d = dir($pPathIn);
 	while (false !== ($entry = $d->read())) {	   
-	   if(	$entry != '.' && $entry != '..' && $entry != 'index.html' && $entry != 'Maintenance.php' 
+	   if(	$entry != '.' && $entry != '..' && $d->path.'/'.$entry != '../index.html' && $entry != 'Maintenance.php' 
 	   		&& $entry != 'Maintenance' && $entry != 'update.sql' && $entry != "DB.php" && $entry != "Mail.php"
 	   		&& $entry != "bdd"
 	   && $entry != ".htaccess") {
@@ -34,7 +34,9 @@ function parcourirDossier($pPathIn,$pPathOut) {
 			if(!is_dir($pPathOut .'/'. $entry)) {
 				mkdir($pPathOut .'/'. $entry);
 			}
-   			parcourirDossier($d->path.'/'.$entry,$pPathOut.'/'. $entry);
+			$lTempIn = $d->path.'/'.$entry;
+			$lTempOut = $pPathOut.'/'. $entry;
+   			parcourirDossier($lTempIn, $lTempOut);
    		} else {
    			$filename = $d->path.'/'.$entry;
 			copy($filename , $pPathOut .'/'. $entry);
